#351 split run_build.py into focused modules
Splits the 1918-line run_build.py into a focused package:

  paths.py         log/report directory layout
  tracking.py      PerformanceTracker + memory/cpu/disk sampling
  display.py       BuildDisplay + SourceProgress + I/O history
  stages.py        worker function + 4 stage runners (run/merge/release/upload)
  orchestrator.py  run_full_build — stage sequencing + post-upload
  run_build.py     thin click CLI shim (preserved as the make build entry point)

Behavior unchanged: manual make run / make merge-all / make release /
make upload-all continue to call the standalone CLIs directly (lightweight,
no orchestrator overhead). make build routes through run_full_build.

Test imports updated to point at the new module locations; monkeypatch
targets updated so existing tests continue to pin the right bindings.

Docstrings stripped across the new modules (kept on click main() since
click consumes it for --help output).
